Update models
Refactored serialisation to support this
